requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
PyGithub==2.3.0
groq==0.9.0
pylint==3.2.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
